+++ packages/cli/src/commands/index.ts
@@ -1,3 +1,4 @@
 export * from './new';
 export * from './config';
 export * from './env';
+export * from './set-lang';