+++ packages/cli/src/project/attributes/env.ts
@@ -23,9 +23,11 @@ export class EnvAttribute implements IProjectAttribute {
 
   async csharp(targetDir: string) {
     const changeCase = await import('change-case');
-    const key = changeCase.pascalCase(this._key, {
-      delimiter: '.'
-    });
+    // Ensures keys like "teams.clientId" are converted to "Teams.ClientId"
+    // It follows the C# convention for environment variables
+    const key = this._key.split('.')
+      .map(part => changeCase.pascalCase(part))
+      .join('.');
 
     return new Compound(new FileJsonSet(targetDir, this._filename, key, this._value));
   }