+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @microsoft/teams.common
 
+## 2.0.0-preview.3
+
+### Patch Changes
+
+- add csharp support to CLI officially, add a2a support via new package, devtools updates
+
 ## 2.0.0-preview.2
 
 ### Patch Changes