+++ tests/cards/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tests/cards
 
+## 0.0.1-preview.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @microsoft/teams.ai@2.0.0-preview.3
+  - @microsoft/teams.apps@2.0.0-preview.3
+  - @microsoft/teams.cards@2.0.0-preview.3
+  - @microsoft/teams.dev@2.0.0-preview.3
+
 ## 0.0.1-preview.2
 
 ### Patch Changes